test: add more fixtures to conftest and use where appropriate
